+++ app/Controllers/configureController.php
@@ -178,7 +178,7 @@ class FreshRSS_configure_Controller extends FreshRSS_ActionController {
 		FreshRSS_View::appendScript(Minz_Url::display('/scripts/draggable.js?' . @filemtime(PUBLIC_PATH . '/scripts/draggable.js')));
 
 		if (Minz_Request::isPost()) {
-			$share = $_POST['share'] ?? null;
+			$share = $_POST['share'] ?? [];
 			if (is_array($share)) {
 				$share = array_filter($share, fn($value, $key): bool =>
 					is_int($key) && is_array($value) &&