+++ lib/lib_rss.php
@@ -1,10 +1,6 @@
 <?php
 declare(strict_types=1);
 
-if (version_compare(PHP_VERSION, FRESHRSS_MIN_PHP_VERSION, '<')) {
-	die(sprintf('FreshRSS error: FreshRSS requires PHP %s+!', FRESHRSS_MIN_PHP_VERSION));
-}
-
 if (!function_exists('mb_strcut')) {
 	function mb_strcut(string $str, int $start, ?int $length = null, string $encoding = 'UTF-8'): string {
 		return substr($str, $start, $length) ?: '';
@@ -348,10 +344,12 @@ function customSimplePie(array $attributes = [], array $curl_options = []): \Sim
 	]);
 	$simplePie->rename_attributes(['id', 'class']);
 	$simplePie->strip_attributes(array_merge($simplePie->strip_attributes, [
-		'autoplay', 'class', 'form', 'formaction',
-		'onload', 'onunload', 'onclick', 'ondblclick', 'onmousedown', 'onmouseup',
-		'onmouseover', 'onmousemove', 'onmouseout', 'onfocus', 'onblur',
-		'onkeypress', 'onkeydown', 'onkeyup', 'onselect', 'onchange', 'seamless', 'sizes', 'srcdoc', 'srcset']));
+		'alink', 'autoplay', 'background', 'bgcolor', 'class', 'form', 'formaction',
+		'link', 'onblur', 'onchange', 'onclick', 'ondblclick', 'onfocus',
+		'onkeydown', 'onkeypress', 'onkeyup', 'onload', 'onmousedown', 'onmousemove',
+		'onmouseout', 'onmouseover', 'onmouseup', 'onselect', 'onunload',
+		'seamless', 'sizes', 'srcdoc', 'srcset', 'text', 'vlink',
+	]));
 	$simplePie->add_attributes([
 		'audio' => ['controls' => 'controls', 'preload' => 'none'],
 		'iframe' => [
@@ -651,11 +649,9 @@ function lazyimg(string $content): string {
 
 /** @return numeric-string */
 function uTimeString(): string {
-	$t = @gettimeofday();
-	$sec = is_numeric($t['sec']) ? (int)$t['sec'] : 0;
-	$usec = is_numeric($t['usec']) ? (int)$t['usec'] : 0;
-	$result = ((string)$sec) . str_pad((string)$usec, 6, '0', STR_PAD_LEFT);
-	return ctype_digit($result) ? $result : '0';
+	$t = gettimeofday();
+	// @phpstan-ignore return.type
+	return ((string)$t['sec']) . str_pad((string)$t['usec'], 6, '0', STR_PAD_LEFT);
 }
 
 function invalidateHttpCache(string $username = ''): bool {