+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 See also [the FreshRSS releases](https://github.com/FreshRSS/FreshRSS/releases).
 
+## 2025-06-02 FreshRSS 1.26.3
+
+* Features
+	* Keep sort and order criteria during navigation [#7585](https://github.com/FreshRSS/FreshRSS/pull/7585)
+	* Add info about `PDO::ATTR_CLIENT_VERSION` (relevant for MySQL / MariaDB with obsolete driver) [#7591](https://github.com/FreshRSS/FreshRSS/pull/7591)
+* Bug fixing
+	* Fix SQL request for user labels with custom sort (affecting PostgreSQL) [#7588](https://github.com/FreshRSS/FreshRSS/pull/7588)
+	* Fix regression for favicon in GReader and Fever APIs [#7573](https://github.com/FreshRSS/FreshRSS/pull/7573)
+	* Fix newest articles (within last second) not shown [#7577](https://github.com/FreshRSS/FreshRSS/pull/7577)
+	* Fix duplicate HTTP header for POST [#7556](https://github.com/FreshRSS/FreshRSS/pull/7556)
+	* Fix important articles on reader view [#7602](https://github.com/FreshRSS/FreshRSS/pull/7602)
+	* Fix remove last share method [#7613](https://github.com/FreshRSS/FreshRSS/pull/7613)
+	* Fix API handling of default category [#7610](https://github.com/FreshRSS/FreshRSS/pull/7610)
+	* Fix user self-deletion [#7626](https://github.com/FreshRSS/FreshRSS/pull/7626)
+	* Move PHP minimum version check [#7560](https://github.com/FreshRSS/FreshRSS/pull/7560)
+* Security
+	* Fix encoding of themes [#7565](https://github.com/FreshRSS/FreshRSS/pull/7565)
+	* Fix `.htaccess.dist` for access to `/scripts/vendor/` [#7598](https://github.com/FreshRSS/FreshRSS/pull/7598)
+* SimplePie
+	* Strip more HTML deprecated styles attributes: `bgcolor, text, background, link, alink, vlink` [#7606](https://github.com/FreshRSS/FreshRSS/pull/7606)
+* UI
+	* Implement loading spinner for marking as favourite/read [#7564](https://github.com/FreshRSS/FreshRSS/pull/7564)
+	* Provide theme class for CSS [#7559](https://github.com/FreshRSS/FreshRSS/pull/7559)
+* Deployment
+	* Use HTTP `Cache-Control: immutable` for some files [#7552](https://github.com/FreshRSS/FreshRSS/pull/7552)
+	* Drop Apache 2.2 (only support Apache 2.4+) [#7561](https://github.com/FreshRSS/FreshRSS/pull/7561)
+* I18n
+	* Improve Indonesian [#7622](https://github.com/FreshRSS/FreshRSS/pull/7622)
+	* Improve Polish [#7587](https://github.com/FreshRSS/FreshRSS/pull/7587)
+* Misc.
+	* Update to PHPMailer 6.10.0 [#7542](https://github.com/FreshRSS/FreshRSS/pull/7542)
+	* Update dev dependencies [#7630](https://github.com/FreshRSS/FreshRSS/pull/7630), [#7631](https://github.com/FreshRSS/FreshRSS/pull/7631), [#7632](https://github.com/FreshRSS/FreshRSS/pull/7632)
+		[#7633](https://github.com/FreshRSS/FreshRSS/pull/7633), [#7634](https://github.com/FreshRSS/FreshRSS/pull/7634), [#7635](https://github.com/FreshRSS/FreshRSS/pull/7635)
+
+
 ## 2025-05-03 FreshRSS 1.26.2
 
 * Features